Runs in x64, add cmake scripts, deprecated old vs projects.

+++ src/server/sv_bot.c
@@ -542,7 +542,7 @@ BotImport_AICast_VisibleFromPos
 */
 qboolean BotImport_AICast_VisibleFromPos(   vec3_t srcpos, int srcnum,
 											vec3_t destpos, int destnum, qboolean updateVisPos ) {	
-	return game->AICast_VisibleFromPos((int)srcpos, srcnum, (int)destpos, destnum, updateVisPos );
+	return game->AICast_VisibleFromPos(srcpos, srcnum, destpos, destnum, updateVisPos );
 }
 
 /*
@@ -551,7 +551,7 @@ BotImport_AICast_CheckAttackAtPos
 ===============
 */
 qboolean BotImport_AICast_CheckAttackAtPos( int entnum, int enemy, vec3_t pos, qboolean ducking, qboolean allowHitWorld ) {	
-	return game->AICast_CheckAttackAtPos(entnum, enemy, (int)pos, ducking, allowHitWorld );
+	return game->AICast_CheckAttackAtPos(entnum, enemy, pos, ducking, allowHitWorld );
 }
 // done.
